Dev: start new file for Explorer create commands
